Refactor Makefile
1. Make file extension consistent
2. Generate dependency files
3. Makefile cleanup
